+++ src/main/groovy/com/github/Lunatix01/SaxonPlugin.groovy
@@ -1,4 +1,4 @@
-package com.github.eerohele
+package com.github.Lunatix01
 
 import org.gradle.api.Project
 import org.gradle.api.Plugin
