+++ src/main/groovy/com/github/Lunatix01/SaxonXsltTask.groovy
@@ -1,6 +1,7 @@
-package com.github.eerohele
+package com.github.Lunatix01
 
-import groovy.util.slurpersupport.GPathResult
+import groovy.xml.XmlSlurper
+import groovy.xml.slurpersupport.GPathResult
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.tasks.Classpath
