+++ src/main/groovy/com/github/Lunatix01/XsltTransformation.groovy
@@ -1,4 +1,4 @@
-package com.github.eerohele
+package com.github.Lunatix01
 
 import net.sf.saxon.Transform
 import org.gradle.workers.WorkAction
