+++ src/main/groovy/com/github/Lunatix01/XsltWorkParameters.groovy
@@ -1,4 +1,4 @@
-package com.github.eerohele
+package com.github.Lunatix01
 
 import org.gradle.api.provider.ListProperty
 import org.gradle.workers.WorkParameters
